requirements.txt
altgraph==0.17
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
idna==2.10
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-server==1.0.4
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lime==0.2.0.1
llvmlite==0.35.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numba==0.52.0
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.11.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
PyFladesk==1.1
Pygments==2.7.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyodbc==4.0.30
pyparsing==2.4.7
PyQt5==5.10
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
shap==0.37.0
sip==4.19.8
six==1.15.0
slicer==0.0.3
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.12.4
tornado==6.0.4
tqdm==4.54.1
traitlets==5.0.4
urllib3==1.25.10
voila==0.2.3
voila-gridstack==0.0.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
XlsxWriter==1.3.7
zipp==3.3.0
